+++ tests/test_pyhgvs.py
@@ -5,7 +5,7 @@ from os.path import abspath
 
 import pyhgvs
 
-from cdot.pyhgvs.pyhgvs_transcript import JSONPyHGVSTranscriptFactory
+from cdot.pyhgvs.pyhgvs_transcript import JSONPyHGVSTranscriptFactory, is_sacgf_pyhgvs_fork
 from .genome import MockGenomeTestFile
 
 
@@ -24,8 +24,13 @@ class TestPyHGVS(unittest.TestCase):
             filename=os.path.join(this_file_dir, 'test_data/grch37.genome'),
             create_data=False)
 
+        sacgf_pyhgvs_fork = is_sacgf_pyhgvs_fork()
+
+        def get_transcript(transcript_id):
+            return factory.get_transcript_grch37(transcript_id, sacgf_pyhgvs_fork=sacgf_pyhgvs_fork)
+
         for hgvs_c, expected_hgvs_g in HGVS_C_TO_G:
-            result = pyhgvs.parse_hgvs_name(hgvs_c, genome, get_transcript=factory.get_transcript_grch37)
+            result = pyhgvs.parse_hgvs_name(hgvs_c, genome, get_transcript=get_transcript)
             name = pyhgvs.HGVSName(expected_hgvs_g)
             expected = (name.chrom, name.start, name.ref_allele, name.alt_allele)
             self.assertEqual(result, expected)